requirements.txt
langchain_openai
langchain
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn
sse-starlette
requests>=2.31.0
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
cassio
groq
langchain-groq